requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
nsfw-detector
pillow
numpy
